refactor(playlist): saca estilos y cabecera de SelectedSongsScreen (2039 -> 1700)
Primer corte del gigante más tocado del repo (PLAN_CALIDAD Fase 1.1). Movimiento
puro: ni un valor de estilo ni una línea de marcado cambian.

- components/playlist/selectedSongsStyles.ts: las 250 líneas de createStyles,
  mismo patrón que gruposStyles.ts y wizardStyles.ts. Sigue siendo factoría
  (depende del esquema de color y del layout) y la pantalla la memoiza igual.
- components/playlist/PlaylistHeaderBar.tsx: la cabecera del resumen (banner de
  coro, cuenta, código de la nube, conmutador de vista). De paso su prop `styles`
  deja de ser `any` y se tipa con SelectedSongsStyles.
- components/playlist/DraggableManualRow.tsx: el wrapper de fila arrastrable.

452 tests en verde; typecheck y lint sin errores y 93 warnings, igual que main.
